Time: 0 ms (100.00%) | Memory: 10.2 MB (24.87%) - LeetSync

+++ 2395-longest-binary-subsequence-less-than-or-equal-to-k/longest-binary-subsequence-less-than-or-equal-to-k.cpp
@@ -8,7 +8,7 @@ public:
         for(int i=1;i<n;i++) pref[i] = pref[i-1]+(s[i]=='0');
         long long ans = pref[n-1];
         long long curr = 0;
-        for(long long i=n-1,j = 0;i>=0 && j < 60 ;i--,j++){
+        for(int i=n-1,j = 0;i>=0 && j < 60 ;i--,j++){
             long long val = n-i;
             curr = (s[i]-'0')*(1LL << j)+curr;
             if(curr <= k) ans = max(ans,(i>0 ? pref[i-1] : 0) +val);
